Разнёс Scenarios по файлам
