+++ src/utils/constants.py
@@ -74,7 +74,7 @@ MAX_ASPECTS_SOLVING_PATH_LEN = 10
 IMAGES_TOLERANCE_PERCENT = 0.02
 IMAGE_TMP_PATH = to_appdata_path('.tmp/tmp.png')
 LOG_FILE_PATH = to_appdata_path('logs/logs.log')
-LINK_GENERATION_MAX_TIME_MS = 10 * 1000
+LINK_GENERATION_MAX_TIME_MS = 1000 * 10  # 10 sec
 
 # Loggers
 MAX_LOG_FILE_SIZE_BYTES = 1024 * 1024 * 5  # 5 Mb
